Move to using queryset all to alleviate test cache
